requirements.txt
numpy==1.19.3
scipy
reportlab
pillow
matplotlib
scikit-learn
